# api/core/openapi.yaml
openapi: 3.0.3
info:
  title: Flight Collaboration Platform Core API
  version: 1.0.0
  description: Internal management and Core business boundary. Core owns business truth.
servers:
  - url: /
paths:
...
  /api/v1/admin/auth/sso/start:
    get:
      summary: Start management-console SSO
      description: Creates a one-time server-side SSO state and redirects the browser to the configured management identity provider (currently enterprise WeCom OAuth; OIDC remains an optional adapter).
      parameters:
        - name: state
          in: query
          required: true
          schema: {type: string, minLength: 16, maxLength: 256}
        - name: redirect_uri
          in: query
          required: true
          schema: {type: string, format: uri}
      responses:
        '302': {description: Redirect to the configured management identity provider}
        '400': {description: Invalid redirect URI or state}
        '503': {description: SSO provider unavailable}
  /api/v1/admin/auth/sso/exchange:
    post:
      summary: Exchange a management SSO authorization code
      description: Consumes the server-side state and authorization code, resolves a provisioned admin identity, and issues a Core audience JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminSSOExchangeRequest'}
      responses:
        '200':
          description: Management session issued
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSSOExchangeResponse'}
        '400': {description: Invalid SSO request}
        '401': {description: Invalid or expired SSO state/code}
        '403': {description: Admin identity is unmapped or inactive}
        '503': {description: SSO provider unavailable}
  /api/v1/admin/auth/me:
    get:
      summary: Get the current management session
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Current management session
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AdminSessionResponse'}
        '401': {description: Missing, invalid, expired, or revoked management session}
  /api/v1/admin/auth/logout:
    post:
      summary: Revoke the current management session
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Management session revoked}
        '401': {description: Missing, invalid, expired, or revoked management session}
...
  /api/v1/personnel:
    post:
      summary: Create a project employee
      operationId: createPersonnel
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/PersonnelWriteRequest'}}}
      responses:
        '201': {description: Project employee created, content: {application/json: {schema: {$ref: '#/components/schemas/ManagementResponse'}}}}
        '400': {description: Invalid management input}
        '403': {description: Principal lacks personnel manage permission}
        '409': {description: Employee conflicts with an existing fact}
    get:
      summary: List personnel visible to the authenticated management principal
      operationId: listPersonnel
      security:
        - bearerAuth: []
      parameters:
        - name: work_state
          in: query
          required: false
          schema: {type: string, enum: [idle, reserved, busy, unavailable]}
        - name: team_public_id
          in: query
          required: false
          schema: {type: string, format: uuid}
        - name: area_public_id
          in: query
          required: false
          schema: {type: string, format: uuid}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200': {description: Paginated personnel list, content: {application/json: {schema: {$ref: '#/components/schemas/PersonnelListResponse'}}}}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks personnel read permission or scope}
...
  /api/v1/assignments:
    get:
      summary: List Assignments visible to the authenticated management principal
      operationId: listAssignments
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema: {type: string, enum: [confirmed, accepted, completed, cancelled]}
        - name: task_public_id
          in: query
          required: false
          schema: {type: string, format: uuid}
        - name: personnel_public_id
          in: query
          required: false
          schema: {type: string, format: uuid}
        - name: page
          in: query
          required: false
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Paginated Assignment list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AssignmentListResponse'}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks Assignment read permission or scope}
        '503': {description: Core repository unavailable}
...
  /api/v1/exceptions:
    get:
      summary: List task-site exceptions reported to Core
      description: Returns exception facts scoped by the authenticated management principal.
      operationId: listTaskExceptions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema: {type: string, enum: [open, acknowledged, resolved, rejected]}
        - name: severity
          in: query
          schema: {type: string, enum: [low, medium, high, critical]}
        - name: task_public_id
          in: query
          schema: {type: string, format: uuid}
        - name: personnel_public_id
          in: query
          schema: {type: string, format: uuid}
        - name: page
          in: query
          schema: {type: integer, minimum: 1, default: 1}
        - name: page_size
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          description: Exception list
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExceptionListResponse'}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks exception read permission or scope}
        '400': {description: Invalid exception filter}
  /api/v1/exceptions/{publicID}:
    patch:
      summary: Update a task exception state
      description: A management principal may acknowledge, resolve or reject an exception within its Core Scope. The transition and audit write are transactional.
      operationId: updateTaskException
      security:
        - bearerAuth: []
      parameters:
        - name: publicID
          in: path
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ExceptionUpdateRequest'}
      responses:
        '200':
          description: Exception state updated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ExceptionUpdateResponse'}
        '400': {description: Invalid exception state or note}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks exception management permission or scope}
        '404': {description: Exception not found in the current scope}
        '409': {description: Exception transition conflicts with its current state}
  /api/v1/reports/overview:
    get:
      summary: Read the scoped operational report overview
      description: Aggregates current Core facts by task, flight, assignment, personnel and exception status. Date filters use flight operating_date.
      operationId: getReportOverview
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          schema: {type: string, format: date}
        - name: to
          in: query
          schema: {type: string, format: date}
      responses:
        '200':
          description: Operational report overview
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ReportOverviewResponse'}
        '400': {description: Invalid date range}
        '401': {description: Missing or invalid JWT}
        '403': {description: Principal lacks analytics read permission or scope}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    flightSourceKey:
      type: apiKey
      in: header
      name: X-Flight-Source-Key
...
  schemas:
    ManagementResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data: {type: object}
        request_id: {type: string}
        trace_id: {type: string}
...
    PersonnelWriteRequest:
      type: object
      required: [employee_no, display_name, team_public_id, position_code, capability_code, password]
      properties:
        employee_no: {type: string, pattern: '^[0-9]{4,12}$'}
        display_name: {type: string, maxLength: 128}
        team_public_id: {type: string, format: uuid}
        position_code: {type: string, pattern: '^[a-z][a-z0-9_-]{1,63}$'}
        capability_code: {type: string, pattern: '^[a-z][a-z0-9_.-]{1,63}$'}
        capabilities: {type: array, minItems: 1, maxItems: 1, items: {type: string, pattern: '^[a-z][a-z0-9_.-]{1,63}$'}}
        password: {type: string, format: password, minLength: 8, maxLength: 128}
...
    AdminSSOExchangeRequest:
      type: object
      required: [state, code, redirect_uri]
      properties:
        state: {type: string, minLength: 16, maxLength: 256}
        code: {type: string, format: password}
        redirect_uri: {type: string, format: uri}
    AdminPrincipal:
      type: object
      required: [public_id, type, role]
      properties:
        public_id: {type: string, format: uuid}
        type: {type: string, enum: [human]}
        role: {type: string, enum: [admin, manager, leader, supervisor]}
    AdminSSOExchangeResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [state, access_token, token_type, expires_in, session_public_id, principal]
          properties:
            state: {type: string, enum: [authenticated]}
            access_token: {type: string, format: password}
            token_type: {type: string, enum: [Bearer]}
            expires_in: {type: integer, format: int64}
            session_public_id: {type: string, format: uuid}
            principal: {$ref: '#/components/schemas/AdminPrincipal'}
        request_id: {type: string}
        trace_id: {type: string}
    AdminSessionResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [session_public_id, expires_at, principal]
          properties:
            session_public_id: {type: string, format: uuid}
            expires_at: {type: string, format: date-time}
            principal: {$ref: '#/components/schemas/AdminPrincipal'}
        request_id: {type: string}
        trace_id: {type: string}
    PersonnelListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/Personnel'}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    Personnel:
      type: object
      required: [public_id, user_public_id, employee_no, display_name, area_public_id, team_public_id, position_code, capability_code, capabilities, work_state, status_version, last_state_changed_at, enabled]
      properties:
        public_id: {type: string, format: uuid}
        user_public_id: {type: string, format: uuid}
        employee_no: {type: string}
        display_name: {type: string}
        area_public_id: {type: string, format: uuid}
        area_name: {type: string}
        team_public_id: {type: string, format: uuid}
        team_name: {type: string}
        position_code: {type: string}
        capability_code: {type: string}
        capabilities: {type: array, items: {type: string}}
        work_state: {type: string, enum: [idle, reserved, busy, unavailable]}
        status_version: {type: integer, format: int64, minimum: 0}
        last_state_changed_at: {type: string, format: date-time}
        unavailable_reason: {type: string}
        enabled: {type: boolean}
    AssignmentListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/Assignment'}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    Assignment:
      type: object
      required: [public_id, task_public_id, flight_public_id, flight_display_no, personnel_public_id, personnel_employee_no, personnel_name, area_public_id, team_public_id, status, status_version, confirmation_id, confirmed_by_public_id, confirmed_at, receipt_status, planned_at]
      properties:
        public_id: {type: string, format: uuid}
        task_public_id: {type: string, format: uuid}
        flight_public_id: {type: string, format: uuid}
        flight_display_no: {type: string}
        personnel_public_id: {type: string, format: uuid}
        personnel_employee_no: {type: string}
        personnel_name: {type: string}
        area_public_id: {type: string, format: uuid}
        team_public_id: {type: string, format: uuid}
        status: {type: string, enum: [confirmed, accepted, completed, cancelled]}
        status_version: {type: integer, format: int64, minimum: 0}
        confirmation_id: {type: string}
        confirmed_by_public_id: {type: string, format: uuid}
        confirmed_at: {type: string, format: date-time}
        receipt_status: {type: string, enum: [pending, received]}
        received_at: {type: string, format: date-time}
        accepted_at: {type: string, format: date-time}
        completed_at: {type: string, format: date-time}
        cancelled_at: {type: string, format: date-time}
        cancel_reason: {type: string}
        planned_at: {type: string, format: date-time}
...
    ExceptionListResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [items, page, page_size, total]
          properties:
            items:
              type: array
              items: {$ref: '#/components/schemas/TaskException'}
            page: {type: integer, minimum: 1}
            page_size: {type: integer, minimum: 1, maximum: 100}
            total: {type: integer, format: int64, minimum: 0}
        request_id: {type: string}
        trace_id: {type: string}
    TaskException:
      type: object
      required: [public_id, task_public_id, flight_public_id, flight_display_no, assignment_public_id, personnel_public_id, personnel_name, area_public_id, team_public_id, category, severity, description, status, reported_by_public_id, reported_at]
      properties:
        public_id: {type: string, format: uuid}
        task_public_id: {type: string, format: uuid}
        flight_public_id: {type: string, format: uuid}
        flight_display_no: {type: string}
        assignment_public_id: {type: string, format: uuid}
        personnel_public_id: {type: string, format: uuid}
        personnel_name: {type: string}
        area_public_id: {type: string, format: uuid}
        team_public_id: {type: string, format: uuid}
        category: {type: string}
        severity: {type: string, enum: [low, medium, high, critical]}
        description: {type: string}
        status: {type: string, enum: [open, acknowledged, resolved, rejected]}
        reported_by_public_id: {type: string, format: uuid}
        reported_at: {type: string, format: date-time}
        resolved_by_public_id: {type: string, format: uuid}
        resolved_at: {type: string, format: date-time}
        resolution_note: {type: string}
    ExceptionUpdateRequest:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [acknowledged, resolved, rejected]}
        resolution_note: {type: string, maxLength: 1024}
    ExceptionUpdateResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [public_id, status]
          properties:
            public_id: {type: string, format: uuid}
            status: {type: string, enum: [acknowledged, resolved, rejected]}
        request_id: {type: string}
        trace_id: {type: string}
    ReportOverviewResponse:
      type: object
      required: [data, request_id, trace_id]
      properties:
        data:
          type: object
          required: [task_counts, flight_counts, assignment_counts, personnel_counts, exception_counts]
          properties:
            from: {type: string, format: date}
            to: {type: string, format: date}
            task_counts: {type: object, additionalProperties: {type: integer, format: int64, minimum: 0}}
            flight_counts: {type: object, additionalProperties: {type: integer, format: int64, minimum: 0}}
            assignment_counts: {type: object, additionalProperties: {type: integer, format: int64, minimum: 0}}
            personnel_counts: {type: object, additionalProperties: {type: integer, format: int64, minimum: 0}}
            exception_counts: {type: object, additionalProperties: {type: integer, format: int64, minimum: 0}}
        request_id: {type: string}
        trace_id: {type: string}
...